requirements.txt
django
djangorestframework>=3.14.0
# djangorestframework-simplejwt
django-storages[google]==1.13.2
pyyaml==6.0
requests==2.28.2
django-cors-headers==3.13.0
# python-decouple
djoser==2.1.0
# psycopg2==2.9.5
# coreapi
psycopg2-binary==2.9.5
django-rest-swagger==2.2.0 
django-environ==0.9.0
google-cloud-secret-manager==2.12.0
twilio==7.16.3
django-filter
google-api-python-client==2.85.0
google-cloud-logging==3.5.0
django-bleach==3.0.1
google-cloud-storage==2.10.0
boto3==1.28.10
botocore==1.31.10
django-redis
apscheduler==3.10.4
Pillow==10.0.1
sendgrid==6.10.0